+++ lib/src/models/entity/estate/estate_property_company_model.dart
@@ -19,9 +19,9 @@ class EstatePropertyCompanyModel extends BaseModuleEntity<String> {
   @JsonKey(name: 'body')
   String? body;
   @JsonKey(name: 'geolocationlatitude')
-  int? geolocationlatitude;
+  double? geolocationlatitude;
   @JsonKey(name: 'geolocationlongitude')
-  int? geolocationlongitude;
+  double? geolocationlongitude;
   @JsonKey(name: 'linkLocationId')
   int? linkLocationId;
   @JsonKey(name: 'keyword')
